refactor: move health-check to middlewares module, drop the obsolete ping-action
