+++ src/utils/runtime.ts
@@ -13,25 +13,3 @@ export const RuntimeFlags = {
     isLocal: false,
     isSystemTask: false,
 };
-
-export function healthCheckEventMiddleware(next: (event: any, context?: any) => Promise<any>) {
-    return async function (event: any, context?: any) {
-        if (event.action === 'ping' || event.Records?.[0]?.cf?.request?.uri === '/health-check') {
-            console.log('pong');
-            return {
-                status: 200,
-                statusDescription: 'OK',
-                body: 'OK',
-                headers: {
-                    'cache-control': [
-                        {
-                            key: 'Cache-Control',
-                            value: 'max-age=0',
-                        },
-                    ],
-                },
-            };
-        }
-        return next(event, context);
-    };
-}
